Code and data files that seem to be mostly working to the current plans - Data Files galIds_*.csv are files sent for adding tags to SNIa hosts   for deep fields - They should have the correct number distribution of SN - The conditional distribution P(SN |M) should be proportional to M, as   shown in the Examples.ipynb notebook.
A problem with the current setup is that the galaxies in the catalog
have a cut for r_ab. This means that we are missing hosts in the catalog
which should contribute some SN, even though they are unlikely
contributors because of low stellar mass. One way to address this is to
do a calculation to figure out a minimum stellar mass from which to do
this calculation using our past model. Then we add in the number of SN
that are missed and will be hostless.  A second method is to get hold of
the original simulation (ie. before the cuts were calculated and try to
do the calculation.) A third method is to move to r_ab rather than
stellar mass and use the calculated distance modulus to deal with the
cuts. This would mean we have to figure out the k-correction which would
be possible due to having the SEDs I think.